made required modification for API to collect details for newsletter
