+++ backend/djangoindia/api/urls/__init__.py
@@ -1,3 +1,4 @@
 from .event import urlpatterns as event_urls
+from .newsletter import urlpatterns as newsletter_urls
 
-urlpatterns = event_urls
+urlpatterns = event_urls + newsletter_urls
