+++ backend/djangoindia/api/urls/event.py
@@ -1,9 +1,8 @@
 from django.urls import path
 
-from djangoindia.api.views.event import EventAPIView, NewsletterSubscriptionAPIView
+from djangoindia.api.views.event import EventAPIView
 # URL conf
 urlpatterns = [
     path("events/", EventAPIView.as_view(), name="list_events"),
     path("events/<uuid:pk>/", EventAPIView.as_view(), name="get_event"),
-    path('newsletter/', NewsletterSubscriptionAPIView.as_view(), name='newsletter-subscription'), #end point for the API to collect details
 ]
