+++ backend/djangoindia/api/views/event.py
@@ -1,11 +1,8 @@
-
 from rest_framework import generics, status
-from rest_framework.exceptions import ValidationError
 from rest_framework.mixins import CreateModelMixin, ListModelMixin, RetrieveModelMixin
 from rest_framework.response import Response
-
-from djangoindia.db.models.event import Event, EventRegistration, NewsletterSubscription
-from djangoindia.api.serializers import EventRegistrationSerializer, EventSerializer, NewsletterSubscriptionSerializer
+from djangoindia.db.models.event import Event, EventRegistration
+from djangoindia.api.serializers.event import EventRegistrationSerializer, EventSerializer
 from djangoindia.bg_tasks.event_registration import registration_confirmation_email_task
 
 
@@ -46,27 +43,6 @@ class EventAPIView(generics.GenericAPIView, ListModelMixin, CreateModelMixin, Re
             registration_confirmation_email_task.delay(recipient_email, event_id)
             return Response(
                 {"message": "Registration successful. Confirmation email sent."},
-                status=status.HTTP_201_CREATED,
-            )
-        except Exception as e:
-            return Response(
-                {"error": str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR
-            )
-
-class NewsletterSubscriptionAPIView(generics.GenericAPIView, ListModelMixin, CreateModelMixin, RetrieveModelMixin):
-    queryset = NewsletterSubscription.objects.all()
-    serializer_class = NewsletterSubscriptionSerializer
-
-    def post(self, request, *args, **kwargs):
-        try:
-            if NewsletterSubscription.objects.filter(email=request.data.get("email")).exists():
-                return Response(
-                    {"error": "This email is already subscribed to the newsletter."},
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-            self.create(request, *args, **kwargs)
-            return Response(
-                {"message": "Subscription successful."},
                 status=status.HTTP_201_CREATED,
             )
         except Exception as e:
