+++ backend/djangoindia/db/models/event.py
@@ -64,10 +64,3 @@ class EventRegistration(BaseModel):
         return (
             f"{self.first_name} {self.last_name} ({self.email}) --- {self.event.name}"
         )
-
-class NewsletterSubscription(BaseModel):
-    name = models.CharField(max_length=100)
-    email = models.EmailField()
-
-    def __str__(self):
-        return self.email
